refactor/feat: se arma carpeta user y se mueve archivo User. Se crea clase estudiante
