+++ src/main/kotlin/ar/edu/unsam/pds/models/Program.kt
@@ -1,6 +1,7 @@
 package ar.edu.unsam.pds.models
 
 import ar.edu.unsam.pds.exceptions.ValidationException
+import ar.edu.unsam.pds.models.user.User
 import jakarta.persistence.*
 import java.io.Serializable
 import java.util.*
