+++ src/main/kotlin/ar/edu/unsam/pds/models/user/Student.kt
@@ -1,0 +1,12 @@
+package ar.edu.unsam.pds.models.user
+
+import ar.edu.unsam.pds.models.Course
+import ar.edu.unsam.pds.models.Program
+
+class Student(name: String,
+              lastName: String,
+              email: String, ) : User(name, lastName, email) {
+
+    var programs: MutableSet<Program> = mutableSetOf()
+
+}
