+++ src/main/kotlin/ar/edu/unsam/pds/models/user/User.kt
@@ -1,20 +1,33 @@
-package ar.edu.unsam.pds.models
+package ar.edu.unsam.pds.models.user
 
-import jakarta.persistence.*
+import ar.edu.unsam.pds.models.Schedule
+import ar.edu.unsam.pds.models.Timestamp
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.FetchType
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.JoinTable
+import jakarta.persistence.ManyToMany
+import jakarta.persistence.Table
 import org.hibernate.annotations.OnDelete
 import org.hibernate.annotations.OnDeleteAction
 import java.io.Serializable
-import java.util.*
+import java.util.UUID
 
-@Entity @Table(name = "APP_USER")
+@Entity
+@Table(name = "APP_USER")
 class User(
     var name: String,
     var lastName: String,
     @Column(unique = true) var email: String,
     var image: String = "",
-    var isAdmin: Boolean = false
+    var isAdmin: Boolean = false,
 ) : Timestamp(), Serializable {
-    @Id @GeneratedValue(strategy = GenerationType.UUID)
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
     lateinit var id: UUID
 
     @ManyToMany(fetch = FetchType.EAGER)
