+++ src/main/kotlin/ar/edu/unsam/pds/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package ar.edu.unsam.pds.repository
 
-import ar.edu.unsam.pds.models.User
+import ar.edu.unsam.pds.models.user.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
