+++ src/test/kotlin/ar/edu/unsam/pds/controllers/EventControllerTest.kt
@@ -4,7 +4,7 @@ import ar.edu.unsam.pds.dto.request.EventRequestDto
 import ar.edu.unsam.pds.dto.request.ScheduleRequestDto
 import ar.edu.unsam.pds.dto.response.EventResponseDto
 import ar.edu.unsam.pds.dto.response.ScheduleResponseDto
-import ar.edu.unsam.pds.models.User
+import ar.edu.unsam.pds.models.user.User
 import ar.edu.unsam.pds.security.models.Principal
 import ar.edu.unsam.pds.services.EventService
 import org.junit.jupiter.api.BeforeEach
